requirements.txt
# ============================================================================
# WordBridge - Python Dependencies
# ============================================================================
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment & Configuration
python-dotenv>=1.0.0

# Async HTTP client
aiohttp>=3.9.0

# Data validation & serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# LLM Client (Hunyuan OpenAI-compatible)
openai>=1.10.0

# Async file I/O
aiofiles>=23.2.1

# Template engine
jinja2>=3.1.3

# System utilities
psutil>=5.9.0

# HTTP client (for sync operations if needed)
requests>=2.31.0

# JSON utilities
orjson>=3.9.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Code quality (optional, for development)
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0

# Performance monitoring (optional)
prometheus-client>=0.19.0

# Production server (optional, alternative to uvicorn)
gunicorn>=21.2.0

